# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: CastleRock Investments — Trading Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /trades:
    get:
      operationId: listTrades
      tags: [trades]
      summary: List trades with optional filters
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
        - { name: assetClass, in: query, schema: { type: ["string","null"] } }
        - { name: symbol, in: query, schema: { type: ["string","null"] } }
        - { name: direction, in: query, schema: { type: ["string","null"] } }
        - { name: strategy, in: query, schema: { type: ["string","null"] } }
        - { name: tag, in: query, schema: { type: ["string","null"] } }
        - { name: dateFrom, in: query, schema: { type: ["string","null"] } }
        - { name: dateTo, in: query, schema: { type: ["string","null"] } }
        - { name: untaggedOnly, in: query, schema: { type: ["boolean","null"] } }
        - { name: page, in: query, schema: { type: ["integer","null"] } }
        - { name: pageSize, in: query, schema: { type: ["integer","null"] } }
        - { name: sortBy, in: query, schema: { type: ["string","null"] } }
        - { name: sortDir, in: query, schema: { type: ["string","null"] } }
      responses:
        "200":
          description: Paginated trade list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeListResponse"
...
  /trades/{id}/journal:
    get:
      operationId: getJournal
      tags: [journals]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Journal entry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journal"
    put:
      operationId: upsertJournal
      tags: [journals]
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JournalInput"
      responses:
        "200":
          description: Upserted journal
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Journal"

  # ── CSV Import ────────────────────────────────────────────────────────────
...
  /insights:
    get:
      operationId: getInsights
      tags: [insights]
      summary: Ranked list of pattern insights
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
        - { name: limit, in: query, schema: { type: ["integer","null"] } }
      responses:
        "200":
          description: Insight list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Insight"
  /insights/weekly-briefing:
    get:
      operationId: getWeeklyBriefing
      tags: [insights]
      summary: Auto-generated weekly performance summary
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
      responses:
        "200":
          description: Weekly briefing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WeeklyBriefing"
  /insights/nudges:
    get:
      operationId: getBehavioralNudges
      tags: [insights]
      summary: Behavioral nudges based on recent activity
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
      responses:
        "200":
          description: Nudge list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Nudge"
  /insights/today:
    get:
      operationId: getTodayInsight
      tags: [insights]
      summary: Single highest-priority insight for today panel
      parameters:
        - { name: accountId, in: query, schema: { type: ["integer","null"] } }
      responses:
        "200":
          description: Today's top insight
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Insight"

  # ── Session Plans ──────────────────────────────────────────────────────────
...
  /share/trades/{id}:
    post:
      operationId: createTradeShareLink
      tags: [share]
      summary: Generate a read-only share link for a trade
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
      responses:
        "200":
          description: Share link token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ShareLink"
  /share/view/{token}:
    get:
      operationId: getSharedTrade
      tags: [share]
      summary: Get shared trade by token (read-only)
      parameters:
        - { name: token, in: path, required: true, schema: { type: string } }
      responses:
        "200":
          description: Shared trade view
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TradeDetail"

  # ── Dashboard ─────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    Trade:
      type: object
      properties:
        id: { type: integer }
        importBatchId: { type: ["integer","null"] }
        accountId: { type: ["integer","null"] }
        importSource: { type: ["string","null"] }
        assetClass: { type: string }
        symbol: { type: string }
        instrumentDescription: { type: ["string","null"] }
        direction: { type: string }
        entryDate: { type: string }
        exitDate: { type: ["string","null"] }
        entryPrice: { type: number }
        exitPrice: { type: ["number","null"] }
        quantity: { type: number }
        pnl: { type: ["number","null"] }
        fees: { type: ["number","null"] }
        stopLoss: { type: ["number","null"] }
        takeProfit: { type: ["number","null"] }
        rMultiple: { type: ["number","null"] }
        tags: { type: ["string","null"] }
        setup: { type: ["string","null"] }
        notes: { type: ["string","null"] }
        session: { type: ["string","null"] }
        economicEventNearby: { type: ["boolean","null"] }
        hasJournal: { type: boolean }
        createdAt: { type: string }
        updatedAt: { type: string }
      required: [id, assetClass, symbol, direction, entryDate, entryPrice, quantity, hasJournal, createdAt, updatedAt]
    TradeDetail:
      type: object
      properties:
        trade:
          $ref: "#/components/schemas/Trade"
        journal:
          $ref: "#/components/schemas/Journal"
      required: [trade]
...
    TradeListResponse:
      type: object
      properties:
        trades:
          type: array
          items:
            $ref: "#/components/schemas/Trade"
        total: { type: integer }
        page: { type: integer }
        pageSize: { type: integer }
      required: [trades, total, page, pageSize]

    # ── Journal ───────────────────────────────────────────────────────────────
    Journal:
      type: object
      properties:
        id: { type: integer }
        tradeId: { type: integer }
        whyEntry: { type: ["string","null"] }
        whyExit: { type: ["string","null"] }
        whyStopLoss: { type: ["string","null"] }
        whyTakeProfit: { type: ["string","null"] }
        mistakes: { type: ["string","null"] }
        marketObservation: { type: ["string","null"] }
        confidenceRating: { type: ["integer","null"] }
        ruleFollowed: { type: ["boolean","null"] }
        tiltState: { type: ["string","null"] }
        executionQualityEntry: { type: ["integer","null"] }
        executionQualityExit: { type: ["integer","null"] }
        executionQualityStop: { type: ["integer","null"] }
        strategyRulesChecked: { type: ["string","null"] }
        updatedAt: { type: string }
      required: [id, tradeId, updatedAt]
    JournalInput:
      type: object
      properties:
        whyEntry: { type: ["string","null"] }
        whyExit: { type: ["string","null"] }
        whyStopLoss: { type: ["string","null"] }
        whyTakeProfit: { type: ["string","null"] }
        mistakes: { type: ["string","null"] }
        marketObservation: { type: ["string","null"] }
        confidenceRating: { type: ["integer","null"] }
        ruleFollowed: { type: ["boolean","null"] }
        tiltState: { type: ["string","null"] }
        executionQualityEntry: { type: ["integer","null"] }
        executionQualityExit: { type: ["integer","null"] }
        executionQualityStop: { type: ["integer","null"] }
        strategyRulesChecked: { type: ["string","null"] }

    # ── Import ────────────────────────────────────────────────────────────────
...
    Insight:
      type: object
      properties:
        id: { type: string }
        category: { type: string }
        priority: { type: integer }
        title: { type: string }
        body: { type: string }
        supportingData:
          type: object
          additionalProperties: true
        isActionable: { type: boolean }
        sampleSize: { type: integer }
      required: [id, category, priority, title, body, isActionable, sampleSize]
    WeeklyBriefing:
      type: object
      properties:
        weekStart: { type: string }
        weekEnd: { type: string }
        tradeCount: { type: integer }
        winRate: { type: number }
        totalPnl: { type: number }
        bestDay: { type: ["string","null"] }
        worstDay: { type: ["string","null"] }
        standoutPattern: { type: ["string","null"] }
        recommendation: { type: string }
        ruleAdherenceRate: { type: ["number","null"] }
        topSymbol: { type: ["string","null"] }
      required: [weekStart, weekEnd, tradeCount, winRate, totalPnl, recommendation]
    Nudge:
      type: object
      properties:
        id: { type: string }
        type: { type: string }
        message: { type: string }
        severity: { type: string }
        actionLabel: { type: ["string","null"] }
        actionRoute: { type: ["string","null"] }
      required: [id, type, message, severity]

    # ── Session Plans ──────────────────────────────────────────────────────────
...
    ShareLink:
      type: object
      properties:
        token: { type: string }
        url: { type: string }
        expiresAt: { type: ["string","null"] }
      required: [token, url]

    # ── Dashboard ─────────────────────────────────────────────────────────────
...